moved T cells creation to its own file
