+++ client/components/GameFunctions/createFunctions/index.js
@@ -3,5 +3,8 @@ import { keyboardControls } from './keyboardControls'
 import { scoreAndStars } from './lifeStars';
 import { NPCCells } from './cells'
 import { epithelialCells, makeEpithelialCell, epithelialCellCollision } from './epithelialCells'
+import { tCells, makeTCell } from './tCells'
 
-export {players, keyboardControls, scoreAndStars, NPCCells, epithelialCells, makeEpithelialCell, epithelialCellCollision }
+export { players, keyboardControls, scoreAndStars, NPCCells, 
+  epithelialCells, makeEpithelialCell, epithelialCellCollision,
+  tCells, makeTCell }
